Run more unit tests on Linux:
- used string16 instead of wstring where needed
- replaced GetTickCount calls with portable functions from base/time.h
- misc GCC cleanups
- enabled following unit tests on Linux:

    browser/autocomplete/history_contents_provider_unittest.cc
    browser/autocomplete/history_url_provider_unittest.cc
    browser/autocomplete/keyword_provider_unittest.cc
    browser/bookmarks/bookmark_html_writer_unittest.cc

Review URL: http://codereview.chromium.org/39335

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11249 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/extensions/user_script_master_unittest.cc
@@ -98,7 +98,7 @@ TEST_F(UserScriptMasterTest, NewScripts) {
   FilePath path = script_dir_.AppendASCII("script.user.js");
 
   const char content[] = "some content";
-  int written = file_util::WriteFile(path, content, sizeof(content));
+  size_t written = file_util::WriteFile(path, content, sizeof(content));
   ASSERT_EQ(written, sizeof(content));
 
   message_loop_.Run();
@@ -111,7 +111,7 @@ TEST_F(UserScriptMasterTest, ExistingScripts) {
   FilePath path = script_dir_.AppendASCII("script.user.js");
 
   const char content[] = "some content";
-  int written = file_util::WriteFile(path, content, sizeof(content));
+  size_t written = file_util::WriteFile(path, content, sizeof(content));
   ASSERT_EQ(written, sizeof(content));
 
   scoped_refptr<UserScriptMaster> master(
